+++ tests/Test.ml
@@ -327,6 +327,10 @@ let tests env =
         | str -> printf "TESTO_TEST is set to %S.\n" str);
     t "xfail" ~expected_outcome:(Should_fail "raises exception on purpose")
       (fun () -> failwith "this exception is expected");
+    t "xfail due to invalid output"
+      ~expected_outcome:(Should_fail "produces incorrect output on purpose")
+      ~checked_output:(Testo.stdout ())
+      (fun () -> print_endline "incorrect output printed by the test");
     t "skipped" ~skipped:"some reason" (fun () ->
         failwith "this shouldn't happen");
     t "broken" ~broken:"this test is super flaky" (fun () ->